Update.
	* locale/findlocale.c: Add casts to avoid warnings.
	* locale/localeinfo.h (LIMAGIC): Add cast to avoid warnings.
	* misc/efgcvt_r.c (fcvt_r): Use ssize_t instead of int and add cast
	to avoid warnings.
	* misc/tsearch.c (const_node): New type.
	(trecurse): Correct casts to avoid warnings.
	(__twalk): Likewise.
	* stdlib/tst-limits.c: Add z modifier to formats for WORD_BIT and
	LONG_BIT.

	* debug/backtrace-tst.c (compare): Add casts to avoid warnings.

+++ locale/localeinfo.h
@@ -33,7 +33,7 @@
 #include <intl/loadinfo.h>	/* For loaded_l10nfile definition.  */
 
 /* Magic number at the beginning of a locale data file for CATEGORY.  */
-#define	LIMAGIC(category)	(0x20000828 ^ (category))
+#define	LIMAGIC(category)	((unsigned int) (0x20000828 ^ (category)))
 
 /* Two special weight constants for the collation data.  */
 #define IGNORE_CHAR	2
